Backend/requirements.txt
blinker==1.4
cryptography==3.2.1
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.7.0
Flask-Cors==3.0.9
Flask-SQLAlchemy==2.4.4
Flask-Mail==0.9.1
Flask-Marshmallow==0.14.0
SQLAlchemy==1.3.20
geojson==2.5.0
GeoAlchemy2==0.8.4
gunicorn==20.0.4
Werkzeug==1.0.1
sentry-sdk==0.19.1
psycopg2-binary==2.8.6
python-dotenv==0.15.0
marshmallow-sqlalchemy==0.24.0
bcrypt==3.2.0
